04-PressureTest/runtime/src/pyproject.toml
[project]
name = "04-pressuretest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13"
dependencies = [
    "bedrock-agentcore>=1.0.4",
    "bedrock-agentcore-starter-toolkit>=0.1.26",
    "boto3>=1.40.59",
    "fastapi>=0.120.0",
    "strands-agents>=1.13.0",
    "psutil>=5.9.0",
]
